Consolidate attention BMM2 quantization helpers
