+++ tests/gpu/torch/kernels/common/attention/test_decode_attention.py
@@ -24,7 +24,7 @@ from modelopt.torch.kernels.common.attention import IS_AVAILABLE as TRITON_KERNE
 
 if TRITON_KERNEL_AVAILABLE:
     from modelopt.torch.kernels.common.attention.decode_attention import attention_decode
-    from modelopt.torch.kernels.quantization.attention.v_qdq import fake_quant_v_onwrite
+    from modelopt.torch.kernels.quantization.attention.bmm2_qdq import fake_quant_v_onwrite
 
 NATIVE_E4M3_AVAILABLE = TRITON_KERNEL_AVAILABLE and torch.cuda.get_device_capability() >= (8, 9)
 requires_native_e4m3 = pytest.mark.skipif(
